[Ionut] - Linux port work
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@460 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: e99f15a0800b5423f88a5138a60492e6bc6ab1ed [formerly 7219053cbad530dfa634bdda672b464660048fa7] [formerly dbbbc3f9633aa149bc67c13f4b4530a80dc6c7ba [formerly 157e7b05e87d37d122a019e56054ab17dbe274e4]]
Former-commit-id: 83827114849269a654b91461111635fce2d9d899 [formerly ea16bffb9cb49fb9b9a6d2135ec00a5c84eeb8e0]
Former-commit-id: f10b25b45ad87d018794e5601e062d829125695e

+++ Source Code/Dynamics/Entities/Particles/ConcreteGenerators/ParticleSphereVelocityGenerator.cpp
@@ -16,10 +16,10 @@ void ParticleSphereVelocityGenerator::generate(const U64 deltaTime,
         phi = Random(-floatPI, floatPI);
         theta = Random(-floatPI, floatPI);
         v = Random(_minVel, _maxVel);
-        r = v * std::sinf(phi);
-        p->_velocity[i].z = v * std::cosf(phi);
-        p->_velocity[i].x = r * std::cosf(theta);
-        p->_velocity[i].y = r * std::sinf(theta);
+        r = v * std::sin(phi);
+        p->_velocity[i].z = v * std::cos(phi);
+        p->_velocity[i].x = r * std::cos(theta);
+        p->_velocity[i].y = r * std::sin(theta);
     }
 }
 };
